+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-28 16:25:51
-// Commit: feat(state): add bi-directional event bus handler for state
+// Updated at: 2026-05-28 18:41:30
+// Commit: style(state): harmonize light and dark theme palette for state
 
 class stateComponentState {
-  final String commitRef = '1552';
-  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
-  final DateTime timestamp = DateTime.parse('2026-05-28T16:25:51+05:00');
+  final String commitRef = '1555';
+  final String lastMessage = 'style(state): harmonize light and dark theme palette for state';
+  final DateTime timestamp = DateTime.parse('2026-05-28T18:41:30+05:00');
 }